component 추출
